scripts/mpv-config: pass -Wl,-Bsymbolic when building libmpv
According to ffmpeg's own documentation* we have to add these linker
flags to libmpv when building it against the static library. waf doesn't
need this for mysterious, unknown reasons (the linker flags there are a
bit different) but passing this works fine on meson. Just add an extra
check to the mpv-config script to enable this if needed. Fixes #208.

*: https://ffmpeg.org/platform.html#Advanced-linking-configuration