scripts/mpv-config: check platform when adding flags
When building libmpv against the static ffmpeg libraries, we have to add
the -Wl,-Bsymbolic linker flags. The problem is that these are only
valid for ELF which doesn't apply to all OSes (namely, macOS). Solve
this by simply checking the output from uname and seeing if it matches a
known ELF platform like linux, bsd, or solaris (who uses this?).
Fixes #216.